refactor: move code into archive package
